Implemented a new test for allelic imbalance
The test is based on the current segment only (instead of the chromosome
chunk) and compares the MAD for the mirrored BAF with the MAD for the
unmirrored. The default value for tau is 1.78 instead of sqrt(3).

A new parameter, 'new.imbalance.test', has been added to the
'ascat.aspcf()' method to control if the new or old test should be used.
The default is to use the new test. Set 'new.imbalance.test=F' to use
the old test.